mma-api/requirements.txt
requests
python-dotenv
numpy
tqdm